boot.bin: source asm and C converted to 32-bit build and executed OK. C main executed OK at 800:0 and jumped to asm function OK as it printed out USD sign

+++ test-code/boot_bin_1-c.c
@@ -6,7 +6,7 @@
 //#include <stdio.h>
 //#include <inttypes.h>
 
-long int maxofthree(long int, long int, long int);
+int maxofthree(int, int, int);
 
 int main() {
     maxofthree(1, -4, -7);
